% Generated by roxygen2 (4.0.2): do not edit by hand
\name{do_binary_random_forest_roc}
\alias{do_binary_random_forest_roc}
\title{do_binary_random_forest_roc}
\usage{
do_binary_random_forest_roc(df, class_field, fields_to_exclude,
  test_fraction = 0.1, numtrees = 1000, node_size = 1,
  variables_to_sample = 3)
}
\arguments{
\item{data}{frame input data}

\item{column}{to use as class labels}

\item{vector}{fields to exclude from the random forest classifier.  should NOT include the class label field.}

\item{number}{fraction of the data frame to sample for out-of-sample test data}

\item{number}{number of trees to have the random forest classifier fit}

\item{number}{minimum node size for the decision trees}

\item{number}{number of variables to sample for each random tree}
}
\value{
list List with fitted model, test classification rate, ROC analysis object, and test error rate for the test data
}
\description{
Calculates a random forest analysis of the specified data frame, with the specified class
label field.  The analysis must be a binary classification, because we also return a ROC curve for
classification accuracy.
}

