% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expected_k_pop}
\alias{expected_k_pop}
\title{expected_k_pop}
\usage{
expected_k_pop(n, theta)
}
\arguments{
\item{n}{population size}

\item{theta}{scaled innovation rate}
}
\value{
expected_k
}
\description{
Calculates the expected number of alleles at a locus in a population, evolving under
a Wright-Fisher infinite-alleles drift model.  Employs the approximation from
Ewens 1972, Equation 4, and performs a simple numerical integration.  The approximation
is good enough to match the value from Mathematica 9/10.
}

