% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_data_path}
\alias{get_data_path}
\title{get_data_path}
\usage{
get_data_path(basedir = "", suffix = "", filename = "", args = "")
}
\arguments{
\item{args}{List of command line arguments (to be processed by commandArgs())}

\item{string}{basedir (defaults to NULL)}

\item{string}{suffix (defaults to NULL)}

\item{string}{filename}
}
\description{
Utility function to help analysis be portable across dev environments and EC2 or
StarCluster execution environments.  If passed a "basedir", it uses that as the
basis of a fully qualified path.  If not, it infers a basedir by first checking
for environment variable "R_DATA_BASEDIR".  If this does not exist, it then checks
for the existence of a variable "data_basedir" in the environment (perhaps from an .rprofile).
If this doesn't exist, it uses getwd() as the basedir.  It then appends the "suffix" if
provided, and returns the canonical form of the directory path for the operating system.
}

