% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_parallel_cores_given_os}
\alias{get_parallel_cores_given_os}
\title{get_parallel_cores_given_os}
\usage{
get_parallel_cores_given_os(dev = TRUE)
}
\arguments{
\item{boolean}{development environment?  defaults to TRUE, and applies to Darwin}
}
\value{
number number of cores to use for parallel processing.  Defaults to all available.
}
\description{
Returns the number of cores to be used for parallel/doMC processing, given the
operating system.  On OSX ("Darwin"), this returns half the number of cores
under the assumption that this is a development environment.  This behavior
can be turned off with the argument dev = FALSE.
}

