% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_parsed_binary_confusion_matrix_stats}
\alias{get_parsed_binary_confusion_matrix_stats}
\title{get_parsed_binary_confusion_matrix_stats}
\usage{
get_parsed_binary_confusion_matrix_stats(cm)
}
\arguments{
\item{confusionMatrix}{Caret package confusionMatrix object, produced typically with test data}
}
\value{
data.frame
}
\description{
Helper function which takes a caret package confusionMatrix object, created by
the confusionMatrix() function, and produces a single-row data frame with common variables
parsed from the object.  These include the overall misclassification rate (1-accuracy), accuracy,
Cohen's kappa, ppv and npv, sensitivity and specificity, and a field which indicates the "positive"
label for those stats which treat one label as a "positive identification".
}

