% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_sorted_variable_importance}
\alias{get_sorted_variable_importance}
\title{get_sorted_variable_importance}
\usage{
get_sorted_variable_importance(model)
}
\arguments{
\item{model}{Fitted and tuned caret model object}
}
\value{
data.frame Sorted and rectified variable importance table
}
\description{
Given a fitted model object from caret's train() method, this utility method
produces a properly sorted data frame with the predictor names in a column
rather than as row.names, renames the columns for use in publications, and sorts
by importance value in descending order.  The resulting data frame produces a
good table in pander() and other methods.
}

