% Generated by roxygen2 (4.0.2): do not edit by hand
\name{list_files_for_data_path}
\alias{list_files_for_data_path}
\title{list_files_for_data_path}
\usage{
list_files_for_data_path(basedir = "", directory = "", pattern = "")
}
\arguments{
\item{character}{Absolute path to a base directory for searching for directories and files}

\item{character}{Relative directory or directory path within which files should be searched}

\item{character}{Pattern for searching for files.}
}
\value{
A character vector containing file names in the specified directories that match the pattern, or "" if none matched.
}
\description{
Lists files for a directory, perhaps given a full base directory specified in the
user's environment or in an environment variable.  The goal is to allow the same
code to be easily run in multiple desktop environments, by multiple developers
who may have different disk directory layouts, and then move that code to servers
and compute clusters, simply by changing one variable which points at the "base"
directory in which files might be found.

If the "basedir" argument is given, either through an absolute path, or more
likely by passing the command line arguments from a script or an environment/startup variable,
then this basedir is prepended to the "directory" variable.  This full absolute path
is then used as the starting point to construct a list of files whose names match the
specified pattern.  This list is constructed using the normal base R function list.files(),
and its return value is passed back to the user, so this function is a drop-in replacement.
}

