% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiclass_rf_replicated_testerror}
\alias{multiclass_rf_replicated_testerror}
\title{multiclass_rf_replicated_testerror}
\usage{
multiclass_rf_replicated_testerror(df, class_field, fields_to_exclude,
  test_fraction = 0.2, numtrees = 500, node_size = 1,
  variables_to_sample = 3, replicates = 100)
}
\arguments{
\item{data}{frame input data}

\item{column}{to use as class labels}

\item{vector}{fields to exclude from the random forest classifier.  should NOT include the class label field.}

\item{number}{fraction of the data frame to sample for out-of-sample test data}

\item{number}{number of trees to have the random forest classifier fit}

\item{number}{minimum node size for the decision trees}

\item{number}{number of variables to sample for each random tree}

\item{number}{replicates of the analysis to be performed}
}
\value{
vector set of prediction accuracy values
}
\description{
Performs a multiclass random forest analysis with M replicates of the entire analysis, including independent selections of
train and test data from the underlying DF, and reports a vector of prediction accuracy values (1 - test error).
}

