% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orderBalancedTree}
\alias{orderBalancedTree}
\title{orderBalancedTree}
\usage{
orderBalancedTree(r, h)
}
\arguments{
\item{r}{Branching factor or breadth of the tree, can be scalar or vector}

\item{h}{Height factor or depth of the tree, can be scalar or vector}
}
\description{
\code{orderBalancedTree} returns the order (number of vertices) of a balanced tree with breadth r and height h.  If the arguments
are vectors, the function returns a vector of orders.
}
\keyword{graph,network}

