% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_multiple_roc}
\alias{plot_multiple_roc}
\title{plot_multiple_roc}
\usage{
plot_multiple_roc(...)
}
\arguments{
\item{multiple}{Output objects from calculate_roc_binary_classifier()}
}
\value{
ggplot2 plot object
}
\description{
Plots multiple ROC curves with diagonal abline, given multiple output objects from calculate_roc_binary_classifier()
}

