% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_multiple_roc_from_list}
\alias{plot_multiple_roc_from_list}
\title{plot_multiple_roc_from_list}
\usage{
plot_multiple_roc_from_list(input_list)
}
\arguments{
\item{multiple}{Output objects from calculate_roc_binary_classifier()}
}
\value{
ggplot2 plot object
}
\description{
Plots multiple ROC curves with diagonal abline, given a list of output objects from calculate_roc_binary_classifier().
Alternate wrapper for the function given that sometimes we'll already have a list of objects, say from a long-running
analysis, and not be passing in individual objects.
}

