% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_roc}
\alias{plot_roc}
\title{plot_roc}
\usage{
plot_roc(roc_obj)
}
\arguments{
\item{list}{Output object from calculate_roc_binary_classifier()}
}
\description{
Plots a single ROC curve with diagonal abline, given output objects from calculate_roc_binary_classifier()
}

