% Generated by roxygen2 (4.0.2): do not edit by hand
\name{random_split_dataset}
\alias{random_split_dataset}
\title{random_split_dataset}
\usage{
random_split_dataset(df, test_fraction)
}
\arguments{
\item{df}{data frame to split}

\item{test_fraction}{}
}
\value{
list List with "train" and "test" elements
}
\description{
Randomly splits a data frame into a training and test set, with a given fraction of rows in the
test data set.
}

