% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stat_ellipse}
\alias{stat_ellipse}
\title{stat_ellipse}
\usage{
stat_ellipse(mapping = NULL, data = NULL, geom = "path",
  position = "identity", ...)
}
\description{
ggplot2 confidence ellipse from https://raw.github.com/JoFrhwld/FAAV/master/r/stat-ellipse.R
}

