% Generated by roxygen2 (4.0.2): do not edit by hand
\name{train_randomforest}
\alias{train_randomforest}
\title{train_randomforest}
\usage{
train_randomforest(df, training_fraction, tuning_grid, training_control,
  exclude)
}
\arguments{
\item{data.frame}{the full data set, without splitting into train/test}

\item{num}{the fraction of the data to use for training}

\item{object}{A tuning grid object of the type used by the caret library (in this case a vector of mtry values)}

\item{object}{A training control object of the type used by the caret library}

\item{character}{A vector of the names of columns to exclude from the analysis}
}
\value{
list A list with the fitted result, the training and test data sets, and the elapsed time
}
\description{
Using the training and tuning function from the caret library, tune a random forest
model on the training fraction of a data frame, after excluding a set of columns.
Tuning occurs given the control and tuning grid objects passed, and the final fitted
object is returned along with the elapsed time, training and test data sets in a list.
}

