% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vertical_dotchart}
\alias{vertical_dotchart}
\title{vertical_dotchart}
\usage{
vertical_dotchart(df, x_var = xvar, x_label = xlabel, y_var = yvar,
  y_label = "Experiment", y_group_var = NULL,
  legend_title = "Experiment Group", sort.by.xvar = TRUE,
  group.by.ygroup = TRUE)
}
\arguments{
\item{df}{Data frame to be plotted}

\item{x_var}{Name of the metric variable to be plotted, as a string}

\item{x_label}{Name of the metric variable as a plot legend string}

\item{y_var}{Name of the categorical variable labeling each row in the data frame}

\item{y_label}{Name of the categorical variable, as a plot legend string}

\item{y_group_var}{Name of a categorical variable which groups the rows}

\item{legend_title}{Title of the legend for groups, defaults to "Experiment Group"}

\item{sort.by.xvar}{Boolean flag to sort rows by the metric variable, defaults to TRUE}

\item{group.by.ygroup}{Boolean flag to sort groups of rows by the grouping variable, defaults to TRUE}
}
\value{
ggplot2 object
}
\description{
Produces a vertical dot chart, with a metric variable giving the length of a horizontal
line segment and a dot colored by a grouping variable.  Each row in the data frame becomes a
row in the dotchart.  The y variable is a label for each row, and rows can be grouped
by a grouping variable.  The plot can be sorted in descending order of the x metric variable,
or grouped by the grouping variable and then sorted.
}

