% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_roc_binary_classifier}
\alias{calculate_roc_binary_classifier}
\title{calculate_roc_binary_classifier}
\usage{
calculate_roc_binary_classifier(fit = NULL, test = NULL,
  label_class = NULL, curve_title = NULL)
}
\arguments{
\item{fit}{Tuned and fitting model from caret's train() function}

\item{test}{Data frame representing test data}

\item{label}{Column name representing the true label in the test set}
}
\value{
list List with a ROC object and AUC
}
\description{
Given a fitted binary classification from caret training and tuning, and a test set,
calculate the ROC curve and AUC, returning a list of both.  Uses the ROCR package
since it can handle the original factors and doesn't need the levels to be turned
numeric first.
}

